expense-splitting-app/backend/requirements.txt
alembic==1.14.0
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cors==1.0.1
cryptography==43.0.3
filelock==3.16.1
Flask==2.2.5
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.0.5
future==1.0.0
gevent==24.2.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==2.1.5
outcome==1.3.0.post0
packaging==24.1
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
requests-file==2.1.0
selenium==4.25.0
setuptools==75.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==1.4.46
stripe==11.3.0
tldextract==5.1.2
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.0.4
wheel @ file:///opt/homebrew/Cellar/python%403.12/3.12.6/libexec/wheel-0.44.0-py3-none-any.whl#sha256=d6ad895f80a568fbc066fd6c52c68c5ef40632c4d764d42af13e28e945f1920b
wsproto==1.2.0
zope.event==5.0
zope.interface==7.0.3
